+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Compatibility:
 * `Kernel#lambda` with no block in a method called with a block raises an exception (#2004, @ssnickolay).
 * Implemented `BigDecimal` as C extension to improve compatibility.
 * Comment lines can be placed between fluent dot now (#2004, @ssnickolay).
+* Implemented `rb_make_exception`.
 
 Performance: